some Basic CRUD for books(library)

+++ src/main/java/com/cgi/library/controller/BookController.java
@@ -1,13 +1,20 @@
 package com.cgi.library.controller;
 
 import com.cgi.library.model.BookDTO;
+import com.cgi.library.model.BookStatus;
 import com.cgi.library.service.BookService;
+import jdk.jfr.ContentType;
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.sql.SQLException;
+import java.time.LocalDate;
+import java.util.Date;
 import java.util.UUID;
 
 @RestController
@@ -26,14 +33,18 @@ public class BookController {
     public ResponseEntity<BookDTO> getBook(@RequestParam(value = "bookId") UUID bookId) {
         return ResponseEntity.ok(bookService.getBook(bookId));
     }
-
     @PostMapping("saveBook")
     public ResponseEntity<String> saveBook(@RequestBody BookDTO book) {
-        return ResponseEntity.ok(String.valueOf(bookService.saveBook(book)));
+        book.setId(UUID.randomUUID());
+        book.setAdded(LocalDate.now());
+        return ResponseEntity.ok("\"" + bookService.saveBook(book) + "\"");
     }
 
     @DeleteMapping("deleteBook")
     public ResponseEntity<String> deleteBook(@RequestParam(value = "bookId") UUID bookId) {
+        if(bookService.getBook(bookId).getStatus() != BookStatus.AVAILABLE) {
+            throw new RuntimeException("Unable to Delete! Object used in Checkout");
+        }
         bookService.deleteBook(bookId);
         return ResponseEntity.ok("");
     }
